91 - Makes the WFOTableWithFilter reload table config when the localStorageKey changes
