+++ packages/orchestrator-ui-components/src/pages/processes/getProcessListTabTypeFromString.ts
@@ -1,0 +1,19 @@
+import { WFOProcessListTabType } from './tabConfig';
+
+export const getProcessListTabTypeFromString = (
+    tabId?: string,
+): WFOProcessListTabType | undefined => {
+    if (!tabId) {
+        return undefined;
+    }
+
+    switch (tabId.toUpperCase()) {
+        case WFOProcessListTabType.ACTIVE.toString():
+            return WFOProcessListTabType.ACTIVE;
+        case WFOProcessListTabType.COMPLETED.toString():
+            return WFOProcessListTabType.COMPLETED;
+
+        default:
+            return undefined;
+    }
+};
