+++ packages/orchestrator-ui-components/src/pages/processes/index.ts
@@ -1,1 +1,4 @@
+export * from './getProcessListTabTypeFromString';
+export * from './tabConfig';
 export * from './WFOProcessListPage';
+export * from './WFOProcessesListSubscriptionsCell';
